+++ src/app/app.module.ts
@@ -12,6 +12,7 @@ import { I18nModule } from '../assets/i18n/i18n.module'
 import { TranslateHttpLoader } from '@ngx-translate/http-loader'
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core'
 import { ProfileModule } from './modules/profile/profile.module'
+import { ToastrModule } from 'ngx-toastr'
 // TODO - pending execution over secure environment
 // import { CookieEncryptionHelper } from './helpers/cookie-encryption.helper'
 
@@ -19,6 +20,8 @@ import localeEs from '@angular/common/locales/es'
 import localeCa from '@angular/common/locales/ca'
 import localeEn from '@angular/common/locales/en'
 import { registerLocaleData } from '@angular/common'
+
+import { MentorLoginComponent } from './modules/mentor/mentor-login/mentor-login.component';
 
 registerLocaleData(localeEs, 'es')
 registerLocaleData(localeCa, 'ca')
@@ -34,6 +37,7 @@ export function HttpLoaderFactory (http: HttpClient): any {
   ],
   bootstrap: [AppComponent],
   imports: [BrowserModule,
+    MentorLoginComponent,
     AppRoutingModule,
     BrowserAnimationsModule,
     CoreModule,
@@ -42,6 +46,7 @@ export function HttpLoaderFactory (http: HttpClient): any {
     ChallengeModule,
     ProfileModule,
     I18nModule,
+    ToastrModule.forRoot(),
     TranslateModule.forRoot({
       defaultLanguage: 'ca',
       loader: {