+++ src/environments/environment.ts
@@ -20,6 +20,7 @@ export const environment = {
   BACKEND_SSO_REGISTER_URL: '/auth/register',
   BACKEND_SSO_VALIDATE_TOKEN_URL: '/tokens/validate',
   BACKEND_GITHUB_VALIDATE_ENDPOINT: '/auth/github/authenticate',
+  BACKEND_LOGOUT_ENDPOINT: '/auth/logout',
   ADMIN_USER: '../assets/dummy/admin-user.json',
   USER_SOLUTION: '../assets/dummy/user-solution.mock.json',
   ITINERARY_ID: 'clpb8t1cc000008k0cg1icvl4',
@@ -32,6 +33,6 @@ export const environment = {
   HTTP_CODE_BAD_REQUEST: 400,
   HTTP_CODE_UNAUTHORIZED: 401,
   HTTP_CODE_FORBIDDEN: 403,
-  GITHUB_CLIENT_ID: 'Ov23lieA4pZVvg3Vn9Vt',
-  GITHUB_REDIRECT_URI: 'http://localhost:4200/ita-challenge/challenges'
+  GITHUB_CLIENT_ID: 'Ov23liMlfGdUlcA569gk',
+  AUTH_REDIRECT_URL: 'http://localhost:4200/ita-challenge/challenges'
 }